requirements.txt
# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion models and transformers
diffusers>=0.25.0
transformers>=4.30.0
accelerate>=0.20.0

# LoRA and parameter-efficient fine-tuning
peft>=0.4.0

# Image processing and utilities
Pillow>=9.0.0
numpy>=1.21.0

# Hugging Face ecosystem
huggingface_hub>=0.15.0
tokenizers>=0.13.0

# Text processing and tokenization
protobuf>=3.20.0
sentencepiece>=0.1.99

# Optional: Progress bars and utilities
tqdm>=4.64.0